Update container install instructions for Trento
The separation of the checks from Wanda into a separate package requires
updates to the container installation.  This allows the user to update
the checks separately from Wanda.  An update to the systemd / RPM
instructions is not required, because Wanda pulls the newly created RPM
as part of its dependencies